fix: activity history scroll problem
