+++ src/app/activity-history/page.tsx
@@ -1,9 +1,7 @@
 import React from 'react';
 import { Metadata } from 'next';
 
-import CPageCard from 'src/components/CPageCard';
-import Transactions from 'src/containers/Transactions';
-import AllStreamCard from 'src/containers/AllStreamCard';
+import ActivityContainer from 'src/containers/ActivityContainer';
 
 export const metadata: Metadata = {
   title: 'Fluxity - Activity History',
@@ -14,19 +12,7 @@ export const metadata: Metadata = {
 };
 
 const ActivityHistory = () => {
-  return (
-    <CPageCard
-      divider
-      scroll
-      borderStatus="borderless"
-      className="py-[15px] px-5 mobile:h-[85vh] h-[100%]"
-      dividerResponsiveClassName="mobile:border-none"
-      title={<h1 className="text-2xl text-midnightBlue mb-[10px] mt-1 pl-2">Activity History</h1>}
-    >
-      <AllStreamCard />
-      <Transactions />
-    </CPageCard>
-  );
+  return <ActivityContainer />;
 };
 
 export default ActivityHistory;
